@@ -2,24 +2,24 @@ package com.adlab.balda.activities;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.InputFilter;
 import android.text.TextWatcher;
 import android.view.View;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.adlab.balda.database.BaldaDataBase;
 import com.adlab.balda.R;
-import com.adlab.balda.model.ActivityPlayer;
 import com.adlab.balda.model.Game;
 import com.adlab.balda.model.GameLab;
 import com.adlab.balda.model.Player;
+import com.google.android.material.textfield.TextInputEditText;
+import com.google.android.material.textfield.TextInputLayout;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class GameSettingsActivity extends AppCompatActivity implements View.OnClickListener{
     private static final String ROW_COUNT = "rowCount";
@@ -32,8 +32,8 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
     private Button buttonStartGame;
     private TextView textViewRowCount;
     private TextView textViewColCount;
-    private EditText editTextInitWord;
-    private TextView textViewNonExistentWord;
+    private TextInputEditText editTextInitWord;
+    private TextInputLayout textInputLayoutInitWord;
 
     private BaldaDataBase dataBase;
 
@@ -61,7 +61,7 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
         textViewRowCount = findViewById(R.id.tv_row_count);
         textViewColCount = findViewById(R.id.tv_column_count);
         editTextInitWord = findViewById(R.id.et_initWord);
-        textViewNonExistentWord = findViewById(R.id.tv_non_existent_word);
+        textInputLayoutInitWord = findViewById(R.id.til);
 
         editTextInitWord.setFilters(new InputFilter[]{new InputFilter.AllCaps()});
         editTextInitWord.addTextChangedListener(new TextWatcher() {
@@ -73,7 +73,7 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
                 word = word.toLowerCase();
                 boolean isWordExist = dataBase.isWordExist(word);
                 if (isWordExist) {
-                    textViewNonExistentWord.setVisibility(View.INVISIBLE);
+                    textInputLayoutInitWord.setError(null);
                     if (word.length() != colCount) {
                         buttonStartGame.setEnabled(false);
                         buttonStartGame.setClickable(false);
@@ -82,9 +82,12 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
                         buttonStartGame.setClickable(true);
                     }
                 } else {
-                    textViewNonExistentWord.setVisibility(View.VISIBLE);
+                    textInputLayoutInitWord.setError(getString(R.string.non_existent_word));
                     buttonStartGame.setEnabled(false);
                     buttonStartGame.setClickable(false);
+                }
+                if (word.isEmpty()) {
+                    textInputLayoutInitWord.setError(getString(R.string.need_enter_word));
                 }
             }
             @Override
@@ -98,8 +101,8 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
         buttonStartGame.setOnClickListener(this);
         buttonGenerateWord.setOnClickListener(this);
 
-        textViewRowCount.setText(rowCount + "");
-        textViewColCount.setText(colCount + "");
+        textViewRowCount.setText(String.valueOf(rowCount));
+        textViewColCount.setText(String.valueOf(colCount));
 
         dataBase = new BaldaDataBase(this);
 
@@ -151,7 +154,7 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
     private void generateRandomWord() {
         String word = dataBase.getRandomWord(colCount);
         editTextInitWord.setText(word);
-        textViewNonExistentWord.setVisibility(View.INVISIBLE);
+        textInputLayoutInitWord.setError(null);
         buttonStartGame.setEnabled(true);
         buttonStartGame.setClickable(true);
     }
@@ -175,7 +178,7 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
             buttonReduceRowCount.setEnabled(true);
             buttonReduceRowCount.setClickable(true);
         }
-        textViewRowCount.setText("" + rowCount);
+        textViewRowCount.setText(String.valueOf(rowCount));
     }
 
     private void reduceRowCount() {
@@ -188,7 +191,7 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
             buttonIncreaseRowCount.setEnabled(true);
             buttonIncreaseRowCount.setClickable(true);
         }
-        textViewRowCount.setText("" + rowCount);
+        textViewRowCount.setText(String.valueOf(rowCount));
     }
 
     private void increaseColumnCount() {
@@ -201,7 +204,7 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
             buttonReduceColCount.setEnabled(true);
             buttonReduceColCount.setClickable(true);
         }
-        textViewColCount.setText("" + colCount);
+        textViewColCount.setText(String.valueOf(colCount));
         String word = editTextInitWord.getText().toString().toLowerCase();
         if (word.length() == colCount && dataBase.isWordExist(word)) {
             buttonStartGame.setEnabled(true);
@@ -221,7 +224,7 @@ public class GameSettingsActivity extends AppCompatActivity implements View.OnCl
             buttonIncreaseColCount.setEnabled(true);
             buttonIncreaseColCount.setClickable(true);
         }
-        textViewColCount.setText("" + colCount);
+        textViewColCount.setText(String.valueOf(colCount));
         String word = editTextInitWord.getText().toString().toLowerCase();
         if (word.length() == colCount && dataBase.isWordExist(word)) {
             buttonStartGame.setEnabled(true);
